+++ src/HiPayPaymentPlugin.php
@@ -135,8 +135,12 @@ class HiPayPaymentPlugin extends Plugin
      */
     public static function getShopwareVersion(): ?string
     {
-        return InstalledVersions::getVersion('shopware/core')
-            ?? InstalledVersions::getVersion('shopware/recovery');
+        $version = null;
+        if (InstalledVersions::isInstalled('shopware/core')) {
+            $version = InstalledVersions::getVersion('shopware/core');
+        }
+
+        return $version ?? exec("cd ../ && php bin/console --version | awk '{print $2}'");
     }
 
     public function install(InstallContext $context): void
@@ -233,6 +237,7 @@ class HiPayPaymentPlugin extends Plugin
             'pluginId' => $this->pluginId,
             'salesChannels' => $this->getSalesChannelIds(),
             'position' => $classname::getPosition(),
+            'technicalName' => $classname::getTechnicalName(),
         ];
 
         /** @var EntityRepository $paymentRepository */