+++ src/PaymentMethod/Paypal.php
@@ -3,6 +3,8 @@
 namespace HiPay\Payment\PaymentMethod;
 
 use HiPay\Fullservice\Data\PaymentProduct;
+use HiPay\Fullservice\Gateway\Request\Order\OrderRequest;
+use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
 
 /**
  * Paypal payment Methods.
@@ -38,5 +40,27 @@ class Paypal extends AbstractPaymentMethod
         ];
 
         return $descriptions[$lang] ?? null;
+    }
+
+    public static function addDefaultCustomFields(): array
+    {
+        return [
+            'merchantPayPalId' => '',
+            'color' => 'gold',
+            'shape' => 'pill',
+            'label' => 'paypal',
+            'height' => '40',
+            'bnpl' => true,
+        ];
+    }
+
+    protected function hydrateHostedFields(OrderRequest $orderRequest, array $payload, AsyncPaymentTransactionStruct $transaction): OrderRequest
+    {
+        if ('paypal' === $orderRequest->payment_product && isset($payload['orderID'])) {
+            $providerData = ['paypal_id' => $payload['orderID']];
+            $orderRequest->provider_data = json_encode($providerData);
+        }
+
+        return $orderRequest;
     }
 }