+++ src/Service/NotificationService.php
@@ -583,7 +583,7 @@ class NotificationService
     /**
      * Add order Transaction capture.
      */
-    private function saveCapture(string $status, ?OrderCaptureEntity $capture, float $amount = null, string $operationId = null, HipayOrderEntity $hipayOrder = null): void
+    private function saveCapture(string $status, ?OrderCaptureEntity $capture, ?float $amount = null, ?string $operationId = null, ?HipayOrderEntity $hipayOrder = null): void
     {
         $context = Context::createDefaultContext();
         if (!$capture) {
@@ -598,7 +598,7 @@ class NotificationService
     /**
      * Add order Transaction refund.
      */
-    private function saveRefund(string $status, ?OrderRefundEntity $refund, float $amount = null, string $operationId = null, HipayOrderEntity $hipayOrder = null): void
+    private function saveRefund(string $status, ?OrderRefundEntity $refund, ?float $amount = null, ?string $operationId = null, ?HipayOrderEntity $hipayOrder = null): void
     {
         $context = Context::createDefaultContext();
         if (!$refund) {