+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Version 2.2.0
+
+- **Add** : Add PayPal v2
+- **Fix** : Fixed getter method to recover Shopware version
+- **Fix** : Fixed technical name for payment methods to prepare Shopware version **6.7**
+
 ## Version 2.1.0
 
 - **Add** : Added new payment means